Fix order service

+++ back_end/order_service/src/main/java/com/pet_care/order_service/dto/response/OrderItemResponse.java
@@ -5,17 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderItemResponse {
-
     String productId;
-
     String productName;
-
     Integer quantity;
-
-    Double price;
+    BigDecimal price;
 }
